refactor: split scraper fetch.py into focused modules (ADR-0036)
Splits the 1711-line fetch.py monolith into three focused modules:

- fetch_tiers.py (532 lines): All three fetch tier implementations
  (llms.txt, content negotiation, Playwright) plus FlareSolverr
  fallback, browser-svc integration, and Playwright proxy helpers.

- fetch_quality.py (364 lines): Content quality assessment, barrier
  detection (bot challenges, CAPTCHAs, login walls), HTML-to-markdown
  conversion, and metadata enrichment.

- fetch.py (370 lines, down from 1711): smart_scrape() orchestrator,
  politeness enforcement, quality degradation, and module-level
  proxy config.

Removes 15 duplicated cache functions from fetch.py (now imported
from cache.py). All test imports updated. Public API unchanged.

82 tests pass; 4 pre-existing failures remain (unrelated domain TTL
tests broken by functools.cache on load_settings).

See ADR-0036 for the full design rationale and considered options.

Closes #240
